move aws loader
