+++ src/loader.ts
@@ -1,5 +1,7 @@
 import { loadPassport } from '@/libs/passport'
+import { awsLoader } from '@/libs/aws'
 
 export async function load() {
     loadPassport()
+    awsLoader()
 }
