+++ src/loaders/index.ts
@@ -1,3 +1,2 @@
-export * from './aws'
 export * from './express'
 export * from './mongoose'
